requirements.txt
django==5.0
python-dotenv
google-generativeai
django-bootstrap5
whitenoise
markdown2
requests==2.31.0  # For WordPress API calls
